- added order_by(sort_order) when geting first state (Initial, Final) - we take in account user set priority if exists more records with seme type - StateManager and state_manager joined to one class - added SK, CZ translation
